+++ examples/akafka/kafka_subscriber.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""This script publishes events to a Kafka topic"""
+"""This script consumes events from a Kafka topic"""
 
 # pylint: disable-all
 
@@ -21,12 +21,12 @@
 import json
 from pathlib import Path
 
-from ghga_service_chassis_lib.kafka import EventProducer, KafkaConfigBase
+from ghga_service_chassis_lib.akafka import EventConsumer, KafkaConfigBase
 
 HERE = Path(__file__).parent.resolve()
 
 TOPIC_NAME = "my_topic"
-SERVICE_NAME = "publisher"
+SERVICE_NAME = "consumer"
 EVENT_TYPE = "counter"
 EVENT_KEY = "count"
 KAFKA_SERVER = "kafka:9092"
@@ -34,23 +34,23 @@ KAFKA_SERVER = "kafka:9092"
 with open(HERE / "message_schema.json", "r") as schema_file:
     EVENT_SCHEMAS = {EVENT_TYPE: json.load(schema_file)}
 
+EXEC_FUNCS = {EVENT_TYPE: print}
+
 CONFIG = KafkaConfigBase(
     service_name=SERVICE_NAME, client_suffix="1", kafka_servers=[KAFKA_SERVER]
 )
 
 
 def run():
-    """Runs publishing process."""
+    """Runs subscribing process."""
 
-    with EventProducer(
-        config=CONFIG, topic_name=TOPIC_NAME, event_schemas=EVENT_SCHEMAS
-    ) as producer:
-        # publish 10 messages:
-        for count in range(0, 10):
-            event_payload = {"count": count}
-            producer.publish(
-                event_type=EVENT_TYPE, event_key=EVENT_KEY, event_payload=event_payload
-            )
+    with EventConsumer(
+        config=CONFIG,
+        topic_names=[TOPIC_NAME],
+        event_schemas=EVENT_SCHEMAS,
+        exec_funcs=EXEC_FUNCS,
+    ) as consumer:
+        consumer.subscribe(run_forever=True)
 
 
 if __name__ == "__main__":
