+++ examples/akafka/message_schema.json
