+++ examples/pub_sub/amqp_publisher.py
