+++ examples/pub_sub/amqp_subscriber.py
