+++ ghga_service_chassis_lib/object_storage_dao.py
@@ -23,7 +23,7 @@ import re
 from dataclasses import dataclass
 from typing import Optional
 
-from .utils import DaoGenericBase, OutOfContextError
+from .utils import DaoGenericBase
 
 
 class ObjectStorageDaoError(RuntimeError):
